[feat]: Moved blockheight module
